- cleaned up code a bit further - changed layout of schedule page - changed a few layouts around to match a coherent theme

+++ app/src/main/java/com/suveraapp/objects/Schedule.java
@@ -3,26 +3,23 @@ package com.suveraapp.objects;
 import android.os.Parcel;
 import android.os.Parcelable;
 
-/**
- * Created by Hibatop on 31/08/2016.
- */
 public class Schedule implements Parcelable {
 
     private long amount; //amount of medication needed to take
     private long time; // time for alarm stored in milliseconds
 
-    public Schedule(int amount, long time){
+    public Schedule(int amount, long time) {
         this.amount = amount;
         this.time = time;
     }
 
-    private Schedule(Parcel in){
+    private Schedule(Parcel in) {
         amount = in.readInt();
         time = in.readLong();
     }
 
-    public static final Parcelable.Creator<Schedule> CREATOR = new Parcelable.Creator<Schedule>(){
-        public Schedule createFromParcel(Parcel in){
+    public static final Parcelable.Creator<Schedule> CREATOR = new Parcelable.Creator<Schedule>() {
+        public Schedule createFromParcel(Parcel in) {
             return new Schedule(in);
         }
 
@@ -36,7 +33,7 @@ public class Schedule implements Parcelable {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(Long amount) {
         this.amount = amount;
     }
